extract song search logic
